% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{BaselineMethod.FEAST}
\alias{BaselineMethod.FEAST}
\title{FEAST}
\usage{
BaselineMethod.FEAST(expr, celltypes, celltype.ngenes, nCores.used)
}
\arguments{
\item{expr}{A gene by cell matrix storing the expression values}

\item{celltypes}{A vector indicating cell types of each cell}

\item{celltype.ngenes}{A named list. Names are cell types, and the values
are number of features selected for that cell type.}

\item{nCores.used}{The number of cores actually used}
}
\value{
A list of each cell type's highly variable genes. The number of HVGs
  are equal to the given number.
}
\description{
FEAST
}
