% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{BaselineMethod.NSforest}
\alias{BaselineMethod.NSforest}
\title{NS-Forest}
\usage{
BaselineMethod.NSforest(expr, celltypes, python.path)
}
\arguments{
\item{expr}{A gene by cell matrix storing the expression values. Should be log-normalized counts.}

\item{celltypes}{A vector indicating cell types of each cell}

\item{python.path}{The path to the python. Need to install \pkg{nsforest} in advance.}
}
\value{
A list of 3-dim arrays. Each array corresponds to a stat. The first dim
is genes, the second dim is cell types, and the last dim is a single subject.
\strong{Note}: the length of the first dim may not equal the total number of genes,
since \emph{negative markers} defined in the \href{https://doi.org/10.1101/gr.275569.121}{paper}
are filtered. Besides some values may be \code{NA}s due to the inconsistency between
\emph{positive markers} of each cell type.
}
\description{
NS-Forest
}
