% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{BaselineMethod.scGeneFit}
\alias{BaselineMethod.scGeneFit}
\title{scGeneFit}
\usage{
BaselineMethod.scGeneFit(expr, celltypes, celltype.ngenes, python.path)
}
\arguments{
\item{expr}{A gene by cell matrix storing the expression values.}

\item{celltypes}{A vector indicating cell types of each cell.}

\item{celltype.ngenes}{A named list. The names are cell types, and the values
are number of features selected for that cell type.}

\item{python.path}{The path to the python. Need to install \pkg{scGeneFit} in
advance.}
}
\value{
A named list. Names are unique cell types. Values are selected
  features for that cell type.
}
\description{
scGeneFit
}
