% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data.info.collect}
\alias{data.info.collect}
\title{data.info.collect}
\usage{
data.info.collect(Y, celltype)
}
\arguments{
\item{Y}{gene expression matrix (normalized to 10k)}

\item{celltype}{cell type info of cells}
}
\value{
All necessary sample level summary statistics
}
\description{
calculate effect size and corresponding variance.
}
\details{
*Note!*
This function is not clean. Some of the calculated statistics are not used
in current manuscript.
Such statistics will be marked with `IGNORE!`
}
