% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data.info.reorganize}
\alias{data.info.reorganize}
\title{Re-organize the data generated by \code{data.info.collect}
Convert the sublists containing necessary stats into arrays}
\usage{
data.info.reorganize(marker.info)
}
\arguments{
\item{marker.info}{The data generated by \code{data.info.collect}.}
}
\value{
A list containing converted assays
}
\description{
Re-organize the data generated by \code{data.info.collect}
Convert the sublists containing necessary stats into arrays
}
