% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_estimation.R
\name{EM_est.param_moment}
\alias{EM_est.param_moment}
\title{Second step to estimate \eqn{\pi} and provide posterior prob}
\usage{
EM_est.param_moment(
  dat.info.sum,
  keep.ix,
  min.cutoff,
  max.cutoff,
  celltype.ix = 2,
  pi_init,
  delta.init,
  tau.init,
  q.init,
  maxiter = 1000,
  tol = 10^-10
)
}
\arguments{
\item{dat.info.sum}{Result from marker_summarization, summary statistics for each sample}

\item{keep.ix}{Index for genes with positive log fold change}

\item{min.cutoff}{Remove samples with extreme small log fold change for robust estimtation}

\item{max.cutoff}{Remove samples with extreme large log fold change for robust estimtation}

\item{celltype.ix}{Cell type index}

\item{pi_init}{Estimated initial value of \eqn{\pi}}

\item{delta.init}{Estimated value from \code{EM_var1_g.indep()}, fixed in this round}

\item{tau.init}{Estimated value from \code{EM_var1_g.indep()}, fixed in this round}

\item{q.init}{Estimated value from \code{EM_var1_g.indep()}, fixed in this round}

\item{maxiter}{Control for EM algorithm}

\item{tol}{Control for EM algorithm}
}
\value{
posterior prob, estimates for parameters
}
\description{
Second step to estimate \eqn{\pi} and provide posterior prob
}
