% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_estimation.R
\name{EM_estimate}
\alias{EM_estimate}
\title{EM algorithm main function}
\usage{
EM_estimate(
  dat.info.sum,
  effect_thres = 0.01,
  maxiter = 1000,
  tol = 10^-3,
  min.cutoff = 0.05,
  max.cutoff = 0.95,
  verbose = FALSE
)
}
\arguments{
\item{dat.info.sum}{Output of function \code{data.info.reorganize()}
contains summarized statistics of each sample}

\item{effect_thres}{Threshold for filtering genes with negative mean (\eqn{m_{gk}} in Eq 3 in manuscirpt)}

\item{maxiter}{Maximum iteration number}

\item{tol}{EM stop control}

\item{min.cutoff}{Remove samples with extreme small log fold change for robust estimtation
quantile 0.05}

\item{max.cutoff}{Remove samples with extreme large log fold change for robust estimtation
quantile 0.95}

\item{verbose}{Whether to print details when the function is running.}
}
\value{
posterior probability and estimates of parameters
}
\description{
EM algorithm main function
}
