% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_estimation.R
\name{EM_var1_g.indep}
\alias{EM_var1_g.indep}
\title{First-round EM for each gene}
\usage{
EM_var1_g.indep(
  delta,
  tau,
  q,
  sigma,
  Y,
  maxiter = 1000,
  tol = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{delta}{Init for log-fold change (m_g in Eq 3 in manuscript)}

\item{tau}{Tau, init for variance between samples}

\item{q}{Init for probability of CTS show DE signal}

\item{sigma}{Variances of gene expression}

\item{Y}{Gene expression}

\item{maxiter}{Maximum iteration number}

\item{tol}{EM stop control}

\item{verbose}{Whether to print details when the function is running.}
}
\value{
estimates of delta  tau and q for each gene
}
\description{
just a case with two mixture normal distribution with same variance
in one normal distribution, the mean is fixed at 0.
}
