% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.expression.matrix}
\alias{get.expression.matrix}
\title{Get the expression matrix}
\usage{
get.expression.matrix(
  sce,
  use.raw = FALSE,
  use.norm.rep = NULL,
  normalize = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{use.raw}{Whether use raw counts data. Default is `FALSE`. If `normalize=FALSE`, the function
returns the raw counts directly. If `normalize=TRUE`, the function returns the
normcounts.}

\item{use.norm.rep}{Which representation in \code{assayNames(sce)} will be used.}

\item{normalize}{Whether normalize the data. Default is `FALSE`. Only valid when `use.raw=TRUE`.}
}
\value{
A dense expression matrix (genes by cells).
}
\description{
Get the expression matrix
}
