% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inverse.log}
\alias{inverse.log}
\title{Inverse the log-transformation}
\usage{
inverse.log(X, log.base = c("natural", 10, 2))
}
\arguments{
\item{X}{a matrix containing the normalized counts (genes by cells).}

\item{log.base}{the base of log-transformation.}
}
\value{
the raw count matrix (genes by cells).
}
\description{
Inverse the log-transformation
}
