% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.raw.counts}
\alias{normalize.raw.counts}
\title{Normalize raw counts
This function normalize (not log-normalize!) the raw counts using a target sum of 1e4.}
\usage{
normalize.raw.counts(X)
}
\arguments{
\item{X}{the raw count matrix (genes by cells).}
}
\value{
a matrix containing the normalized counts.
}
\description{
Normalize raw counts
This function normalize (not log-normalize!) the raw counts using a target sum of 1e4.
}
