% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{runBaselineMethod}
\alias{runBaselineMethod}
\title{A wrapper function of classic differential expression tests and feature
selection methods.}
\usage{
runBaselineMethod(
  sce,
  use.norm.rep = NULL,
  subject.rep = "subject",
  celltype.rep = "celltype",
  per.subject = TRUE,
  method = c("wilcox", "twelch", "DEseq2", "NSforest", "FEAST", "scGeneFit"),
  celltype.ngenes = NULL,
  numCores = NULL,
  python.path = NULL
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object. Should contain normalized
count matrix, subject and cell type info.}

\item{use.norm.rep}{Which count matrix should be used. Default is the matrix
accessed by \code{normcounts()}.}

\item{subject.rep}{Column name for the subject info.}

\item{celltype.rep}{Column name for the cell type info.}

\item{per.subject}{If \code{TRUE}, perform DE analysis for each sample.
Otherwise perform DE analysis for all cells.}

\item{method}{A string indicating which DE method to use.}

\item{celltype.ngenes}{A named list. The names are cell types, and the values
are number of features selected for that cell type.}

\item{numCores}{Number of cores to use. Default is \code{NULL}.}

\item{python.path}{The path to the python. Need to install \pkg{nsforest} in
advance.}
}
\value{
A list of 3-dim arrays. Each array is corresponding to a stat. The
  first dim is for genes, the second dim is for cell types, and the last dim
  is for subjects.
}
\description{
Currently, DE tests implemented are:
1. Wilcoxon test
2. t-test
3. ZINB-WaVE + DESeq2
These tests are aimed to find over-expressed DE genes.
FS methods implemented are:
1. NS-Forest
2. FEAST
3. scGeneFit
For each cell type, these methods select a predefined number of markers.
}
\examples{
# load the simulated data
data(sim.sce)

# normalize
sim.sce <- scater::logNormCounts(sim.sce, log=FALSE)

# run the Wilcoxon test per subject
res <- runBaselineMethod(sim.sce,
                         subject.rep='subject', celltype.rep='celltype',
                         per.subject=TRUE, method='wilcox', numCores=2)

}
\references{
Risso, Davide, et al. "A general and flexible method for signal
extraction from single-cell RNA-seq data." Nature communications 9.1 (2018):
284.

Love, Michael, Simon Anders, and Wolfgang Huber. "Differential
analysis of count data–the DESeq2 package." Genome Biol 15.550 (2014):
10-1186.

Aevermann, Brian, et al. "A machine learning method for the
discovery of minimum marker gene combinations for cell type identification
from single-cell RNA sequencing." Genome research 31.10 (2021): 1767-1780.

Su, Kenong, Tianwei Yu, and Hao Wu. "Accurate feature selection
improves single-cell RNA-seq cell clustering." Briefings in bioinformatics
22.5 (2021): bbab034.

Dumitrascu, Bianca, et al. "Optimal marker gene selection for
cell type discrimination in single cell analyses." Nature communications 12.1
(2021): 1186.
}
