% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCTS.R
\name{scCTS}
\alias{scCTS}
\title{The main function of scCTS. It performs DE analysis on normalized (not
log-normalized) count matrix.}
\usage{
scCTS(
  sce,
  use.raw = FALSE,
  use.norm.rep = NULL,
  subject.rep = "subject",
  celltype.rep = "celltype",
  log.input = FALSE,
  log.base = 2,
  effect_thres = 0.01,
  maxiter = 1000,
  tol = 0.001,
  min.cutoff = 0.05,
  max.cutoff = 0.95,
  numCores = NULL,
  verbose = FALSE
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object. Should contain normalized
count matrix, subject and cell type info.}

\item{use.raw}{Whether to use the raw counts. If \code{TRUE}, the raw counts
are then normalized by \pkg{scCTS}.}

\item{use.norm.rep}{Which count matrix should be used. Default is the matrix
accessed by \code{normcounts()}.}

\item{subject.rep}{The name of column that stores subject labels of cells in
\code{colData} slot.}

\item{celltype.rep}{The name of column that stores cell type labels in
\code{colData} slot.}

\item{log.input}{Whether the input expression matrix is log-transformed.}

\item{log.base}{The base of log-transformation.}

\item{effect_thres}{Threshold for filtering genes with negative mean
(\eqn{m_{gk}} in Eq 3 in manuscript)}

\item{maxiter}{Maximum iteration number}

\item{tol}{EM stop control}

\item{min.cutoff}{Remove samples with extreme small log fold change for
robust estimation. Default is quantile 0.05.}

\item{max.cutoff}{Remove samples with extreme large log fold change for
robust estimation. Default is quantile 0.95.}

\item{numCores}{Number of cores to use. Default is the number of all possible
cores minus 1.}

\item{verbose}{Whether to print details when the function is running.}
}
\value{
A list containing lists for every cell type. Each list contains
  posterior probability and estimates of parameters for a cell type.
}
\description{
The main function of scCTS. It performs DE analysis on normalized (not
log-normalized) count matrix.
}
\examples{
# load the simulated data
data(sim.sce)

# run scCTS
res <- scCTS(sim.sce, use.raw = TRUE,
             subject.rep='subject', celltype.rep='celltype',
             numCores=2)

}
