% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.parallel.computation}
\alias{set.parallel.computation}
\title{Set parallel computation}
\usage{
set.parallel.computation(numCores = NULL)
}
\arguments{
\item{numCores}{the number of cores to use. Default is \code{NULL}, which
means using all but one possible cores.}
}
\value{
the number of actually used cores.
}
\description{
Set parallel computation
}
