% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stratified.sampling}
\alias{stratified.sampling}
\title{Perform a stratified sampling on the cells from specified cell types}
\usage{
stratified.sampling(sce, sub.rep, ct.rep, fraction, subset.cts)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{sub.rep}{The name of subject labels.}

\item{ct.rep}{The name of cell type labels.}

\item{fraction}{The sampling fraction between 0 and 1.}

\item{subset.cts}{Cell types that require sampling. Must be a subset of \code{unique(colData(sce)[[ct.rep]])}.}
}
\value{
the sampling \code{SingleCellExperiment} object.
}
\description{
Perform a stratified sampling on the cells from specified cell types
}
