% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{BaselineMethod.DESeq2}
\alias{BaselineMethod.DESeq2}
\title{DESeq2}
\usage{
BaselineMethod.DESeq2(
  expr,
  celltypes,
  nCores.used,
  subjects = NULL,
  include.subject = T
)
}
\arguments{
\item{expr}{A gene by cell matrix storing the expression values}

\item{celltypes}{A vector indicating cell types of each cell}

\item{nCores.used}{The number of cores actually used}

\item{subjects}{A vector indicating subjects of each cell}

\item{include.subject}{Whether to include subject factor in the formula}
}
\value{
A list of 3-dim arrays. Each array corresponds to a stat. The first dim
is genes, the second dim is cell types, and the last dim is a single subject.
}
\description{
DESeq2
}
