% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{BaselineMethod.NSforest}
\alias{BaselineMethod.NSforest}
\title{NS-Forest}
\usage{
BaselineMethod.NSforest(
  expr,
  celltypes,
  celltype.ngenes,
  python.path,
  nCores.used = NULL
)
}
\arguments{
\item{expr}{A gene by cell matrix storing the expression values}

\item{celltypes}{A vector indicating cell types of each cell}

\item{celltype.ngenes}{A named list. The names are cell types, and the values
are number of features selected for that cell type.}

\item{python.path}{The path to the python.}

\item{nCores.used}{The number of cores actually used}
}
\value{
A named list. Names are unique cell types. Values are selected
  features for that cell type.
}
\description{
Running this function will create a folder ./NSForest_outputs/ in the current
working directory. For each cell type, the genes should be sorted first by
binary scores and then by feature importances from the random forest model.
The number of actually selected features may be less than the given number of
selected features, since \emph{negative markers} defined in the
\href{https://doi.org/10.1101/gr.275569.121}{paper} are filtered.
}
