% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{runBaselineMethod}
\alias{runBaselineMethod}
\title{A wrapper function of the classic methods to find over-expressed DE genes}
\usage{
runBaselineMethod(
  sce,
  use.raw = FALSE,
  use.norm.rep = NULL,
  normalize = FALSE,
  log.input = FALSE,
  log.base = 2,
  subject.rep = "subject",
  celltype.rep = "celltype",
  per.subject = TRUE,
  method = c("wilcox", "twelch", "DEseq"),
  numCores = NULL
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object. Should contain normalized count matrix,
subject and cell type info.}

\item{use.raw}{Whether to use the raw count matrix, which is normalized internally.}

\item{use.norm.rep}{Which count matrix should be used. Default is the matrix
accessed by \code{normcounts()}.}

\item{normalize}{Whether to normalize the raw counts. Only relevant when \code{use.raw=TRUE}.}

\item{log.input}{Whether the input expression matrix is log-transformed.}

\item{log.base}{The base of log-transformation.}

\item{subject.rep}{Column name for the subject info.}

\item{celltype.rep}{Column name for the cell type info.}

\item{per.subject}{If \code{TRUE}, perform DE analysis for each sample.
Otherwise perform DE analysis for all cells.}

\item{method}{A string indicating which DE method to use.}

\item{numCores}{Number of cores to use. Default is \code{NULL}.}
}
\value{
A list of 3-dim arrays. Each array is corresponding to a stat. The first dim
is for genes, the second dim is for cell types, and the last dim is for subjects.
}
\description{
A wrapper function of the classic methods to find over-expressed DE genes
}
\examples{
# load the simulated data
data(sim.sce)

# run the Wilcoxon test per subject
res <- runBaselineMethod(sim.sce, use.raw = TRUE, normalize=TRUE,
                         subject.rep='subject', celltype.rep='celltype',
                         per.subject=TRUE, method='wilcox', numCores=2)

}
